+++ CHANGES.rst
@@ -3,7 +3,7 @@ Changelog (nionui)
 
 0.4.0 (UNRELEASED)
 ------------------
-- Add show_ok_dialog to application.
+- Add show_ok_dialog and show_ok_cancel_dialog methods to application class.
 - Add run_ui to application for running simple declarative windows.
 - Add functions for opening file dialogs without a window.
 - Add support for adjusting menus at the application level.